Use long ints in sampler to avoid overflow.

+++ muon/sampler.cc
@@ -47,9 +47,9 @@ bool Sampler::NextSample(float &x, float &y) {
   return true;
 }
 
-int Sampler::TotalSamples() const { return total_samples_; }
+long int Sampler::TotalSamples() const { return total_samples_; }
 
-int Sampler::RequestedSamples() const { return samples_; }
+long int Sampler::RequestedSamples() const { return samples_; }
 
 float Sampler::Progress() const {
   return samples_ / static_cast<float>(total_samples_);
